int->long try 2

+++ src/main/java/org/example/Main.java
@@ -44,10 +44,10 @@ public class Main {
                     } else if (x > 12) {
                         // 13! overflows a standard 32-bit int
                         System.out.println("Warning: Factorial of " + x + " will overflow 'int' and give an incorrect result.");
-                        int result = calculatorService.fact(x);
+                        long result = calculatorService.fact(x);
                         System.out.println("Factorial Result (Overflowed): " + result);
                     } else {
-                        int result = calculatorService.fact(x);
+                        long result = calculatorService.fact(x);
                         System.out.println("Factorial Result: " + result);
                     }
                     break;
